Port plugins/plugins_3rdparty to Qt6 — full UGENE build green
- Mass-convert QRegExp -> QRegularExpression (~150+ files across plugins).
- Rewrite indexIn/cap/matchedLength patterns to QRegularExpressionMatch
  in ClustalO/W, MAFFT, MrBayes, Fastqc, Bowtie, BowtieTask, BlastDB,
  QDDocument, RemoteBLASTTask, ExternalToolValidateTask, TaskTests
  flagsFromString, SchemeWrapper-like patterns.
- Convert QString::contains/split/replace(QRegExp,...) call sites by
  converting the QRegExp argument to QRegularExpression.
- Fix Qt::CTRL+key shortcut chains -> Qt::CTRL|key in enzymes/MaConsensus/
  MainWindow/MDIManager etc (operator+ deleted for QFlags in Qt6).
- Qt6 API: QString(int) -> QString::number; setBackgroundColor -> set-
  Background; QListWidgetItem; QString::leftRef -> QStringView::left;
  QApplication::desktop -> QGuiApplication::primaryScreen with QScreen;
  setValue<T> -> QVariant::fromValue<T>; QMutex::Recursive/NonRecursive
  -> QRecursiveMutex / default; QMutableMapIterator over QMultiMap ->
  QMutableMultiMapIterator; ParseResult -> bool() cast for DomDocument;
  QFileInfo from QString constructor; QString::arg(QFlags) cast to int;
  qBound type-explicit literals; brace-init QPoint disambiguation.
- BALL: strip C++17-incompatible dynamic exception specs (throw(...))
  multi-line aware.
- phylip: comment out '#define Const const' which collides with Qt6
  qiterable.h template parameter; fix bogus filename==NULL on QString.
- biostruct3d_view: link Qt6::OpenGL/OpenGLWidgets; move ui-level slot
  connect (sl_setShadingLevel) out of .ui into C++ (uic Qt6 generates
  typed connect that needs the slot on the form class).
- ugeneui StyleFactory::nativeEventFilter long* -> qintptr*.
- ExternalToolRegistry: versionRegExp QRegExp -> QRegularExpression and
  propagate through dependents.
- QElapsedTimer explicit includes (Qt6 no longer pulls it transitively).
- QActionGroup, QFile, QGuiApplication, QScreen explicit includes.
- TestViewReporter: indexIn->indexOf, setMinimal->setPatternOptions(
  InvertedGreedinessOption).
- Misc fixes: kalign 'if (prefixToken>0)' typo unmasked by Qt6 strict
  typing; cufflinks 0!=ptr -> bool(ptr).

+++ src/plugins/GUITestBase/src/tests/PosteriorActions.cpp
@@ -45,7 +45,7 @@ namespace U2 {
 namespace GUITest_posterior_actions {
 
 POSTERIOR_ACTION_DEFINITION(post_action_0000) {
-    uiLog.trace(QString("post_action_0000: next keyboard modifiers are pressed after the test: %1").arg(QGuiApplication::queryKeyboardModifiers()));
+    uiLog.trace(QString("post_action_0000: next keyboard modifiers are pressed after the test: %1").arg((quint32)QGuiApplication::queryKeyboardModifiers()));
     GTMouseDriver::releasePressedButtons();
     GTKeyboardDriver::releasePressedKeys();
 }
